Remove mutual reference at RenderEffect
