+++ src/Tizen.NUI/src/internal/Interop/Interop.BackgroundBlurEffect.cs
@@ -1,0 +1,11 @@
+namespace Tizen.NUI
+{
+    internal static partial class Interop
+    {
+        internal static partial class BackgroundBlurEffect
+        {
+            [global::System.Runtime.InteropServices.DllImport(NDalicPINVOKE.Lib, EntryPoint = "CSharp_Dali_BackgroundBlurEffect_New__SWIG_1")]
+            public static extern global::System.IntPtr New(uint pixelRadius);
+        }
+    }
+}
