--- src/Tizen.NUI/src/internal/Interop/Interop.RenderEffect.cs
@@ -1,17 +1,0 @@
-namespace Tizen.NUI
-{
-    internal static partial class Interop
-    {
-        internal static partial class BackgroundBlurEffect
-        {
-            [global::System.Runtime.InteropServices.DllImport(NDalicPINVOKE.Lib, EntryPoint = "CSharp_Dali_BackgroundBlurEffect_New__SWIG_0")]
-            public static extern global::System.IntPtr New();
-
-            [global::System.Runtime.InteropServices.DllImport(NDalicPINVOKE.Lib, EntryPoint = "CSharp_Dali_BackgroundBlurEffect_New__SWIG_1")]
-            public static extern global::System.IntPtr New(uint pixelRadius);
-
-            [global::System.Runtime.InteropServices.DllImport(NDalicPINVOKE.Lib, EntryPoint = "CSharp_Dali_delete_BackgroundBlurEffect")]
-            public static extern void Delete(global::System.Runtime.InteropServices.HandleRef self);
-        }
-    }
-}
